Update autho on some files

+++ controllers/circleMembers.controller.js
@@ -1,4 +1,5 @@
 const CircleMembers = require("../models/CircleMembers");
+const Circles = require('../models/Circles')
 
 // GET request here
 async function getAllCircleMembers(req,res) {
@@ -17,11 +18,30 @@ async function getAllCircleMembers(req,res) {
 async function createCircleMembers(req,res) {
   try {
     // TODO: add authorization here--⚠️⚠️
+    // get req info
+    const user = req.user
+    console.log(`user:`,user);
+    // check if req user role is admin
+    const userRole = user.role
+    console.log(`userRole`, userRole);
+    // check if req user role is circle creator
+    const circle = await Circles.findByPk(req.body.circleId)
+    if(!circle) throw `no circle la stupid`
+
+    const circleCreator = circle.creatorId
+
+    if(userRole !== "admin" && circleCreator !== user.id){
+      console.log(`checkpoint 2`);
+      throw `Access denied: This not your circle la`
+    }
+    // if they not, they cannot create member
+    // if they are, they can create the member (invite member)
 
     const createdUser = await CircleMembers.create({
         ...req.body,
     })
     res.json(createdUser)
+    // res.json(`test`)
   } catch (error) {
     res.status(500).json({ error: error });
   }
@@ -33,6 +53,25 @@ async function updateCircleMembers(req,res) {
   try {
     // TODO: add authorization here--⚠️⚠️
     // if(role !== "creator"|| "admin"){}
+    // check if req user is admin or circle-admin
+    const user = req.user
+    console.log(`user:`,user);
+    // check if req user role is admin
+    const userRole = user.role
+    console.log(`userRole`, userRole);
+    // check if req user role is circle creator
+    const circle = await Circles.findByPk(req.body.circleId)
+    if(!circle) throw `no circle la stupid`
+
+    const circleCreator = circle.creatorId
+
+    if(userRole !== "admin" && circleCreator !== user.id){
+      console.log(`checkpoint 2`);
+      throw `Access denied: This not your circle la`
+    }
+    // if not, deny access
+    // if is, update circle member
+    // 
 
     const updatedCircleMember = await CircleMembers.update({
         ...req.body
@@ -56,6 +95,22 @@ async function deleteCircleMembers(req,res) {
     // if(role !== "admin" || "creator"){}
     // else if (user !== "user"){}
     // else{}
+    const user = req.user
+    console.log(`user:`,user);
+    // check if req user role is admin
+    const userRole = user.role
+    console.log(`userRole`, userRole);
+    // check if req user role is circle creator
+    const circle = await Circles.findByPk(req.body.circleId)
+    if(!circle) throw `no circle la stupid`
+
+    const circleCreator = circle.creatorId
+
+    if(userRole !== "admin" && circleCreator !== user.id){
+      console.log(`checkpoint 2`);
+      throw `Access denied: This not your circle la`
+    }
+
     const deletedCircleMember = await CircleMembers.destroy({
         where:{
             id: parseInt(req.params.circleMemberId)
