+++ controllers/circles.controller.js
@@ -4,12 +4,17 @@ const CircleMembers = require("../models/CircleMembers");
 async function getAllCircle(req, res) {
   try {
     // TODO: add authorization here--⚠️⚠️
+    if(req.user.role !== "admin"){
+      const circles = await Circle.findAll({
+        where:{
+          creatorId : req.user.id
+        }
+      });
+      return res.json(circles)
+    }
 
-    //⚠️⚠️to be updated
-   
       const circles = await Circle.findAll();
-      res.json(circles);
-    
+      res.json(circles); 
     }
    catch (error) {
     res.status(500).json({ error: error });
@@ -30,6 +35,10 @@ async function updateCircle(req, res) {
   // TODO: add authorization here--⚠️⚠️
   
     try {
+      const circle = await Circle.findByPk(parseInt(req.params.circleId))
+
+      if( circle.creatorId !== req.user.id && req.user.role !== "admin") throw `Restricted access: Cannot update other user circle`
+
       const updatedCircle = await Circle.update(req.body, {
         where: {
           id: parseInt(req.params.circleId),
@@ -45,6 +54,9 @@ async function updateCircle(req, res) {
 async function deleteCircle(req, res) {
     try {
   // TODO: add authorization here--⚠️⚠️
+  const circle = await Circle.findByPk(parseInt(req.params.circleId))
+
+  if( circle.creatorId !== req.user.id && req.user.role !== "admin") throw `Restricted access: Cannot delete other user circle`
         
         const deletedCircle = await Circle.destroy({
             where:{
